# openapi/bundles/00-graymatter-core.yaml
...
openapi: 3.0.3
info:
  title: GrayMatter Lite API
  version: 0.4.0
  description: Canonical ThorAPI source for the local GrayMatter domain. Generated
    audit, ownership, CRUD, repository, service, controller, and client surfaces are
    derived from this document; generated source is never edited directly.
  contact:
    email: info@valkyrlabs.com
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
- url: "/v1"
  description: GrayMatter Lite local server
...
security:
- basicAuth: []
paths:
...
  "/MemoryEntry/write":
    post:
      tags:
      - LocalRuntimeOperations
      operationId: writeMemoryEntry
      x-thorapi-nonCrud: true
      summary: Write a durable memory through the local compatibility contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateMemoryEntryRequest"
      responses:
        '200':
          description: Persisted memory
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MemoryEntry"
  "/MemoryEntry/read":
    post:
      tags:
      - LocalRuntimeOperations
      operationId: readMemoryEntries
      x-thorapi-nonCrud: true
      summary: Read memories through the compatibility query envelope
      requestBody:
        required: false
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/QueryMemoryEntryRequest"
      responses:
        '200':
          description: Matching memories
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/MemoryEntry"
  "/MemoryEntry/query":
    post:
      tags:
      - LocalRuntimeOperations
      operationId: queryMemoryEntries
      x-thorapi-nonCrud: true
      summary: Query owner-visible memories
      requestBody:
        required: false
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/QueryMemoryEntryRequest"
      responses:
        '200':
          description: Query result envelope
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/MemoryQueryResponse"
...
  "/memory/semantic-index/search":
    post:
      summary: Search owned memories with keyword, vector, or hybrid ranking
      tags:
      - LocalRuntimeOperations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                limit:
                  type: integer
                  minimum: 1
                  maximum: 100
                retrievalMode:
                  type: string
                  enum:
                  - KEYWORD
                  - VECTOR
                  - HYBRID
                  - SCHEMA_FILTERED
                  - RECENCY_BIASED
                type:
                  type: string
                source:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                filters:
                  type: object
                  description: Lite accepts only type, source/sourceChannel, and tags;
                    other keys are rejected
                  additionalProperties: false
                  properties:
                    type:
                      type: string
                    source:
                      type: string
                    sourceChannel:
                      type: string
                    tags:
                      type: array
                      items:
                        type: string
              required:
              - query
      responses:
        '200':
          description: Owner-scoped ranked memory hits with lexical/vector scores
            and index profile
      operationId: graymatterLiteMemorySemanticIndexSearchPost
      x-thorapi-nonCrud: true
  "/memory/reindex":
    post:
      summary: Rebuild the authenticated principal memory vector projection
      tags:
      - LocalRuntimeOperations
      responses:
        '200':
          description: Indexed, semantic-indexed, and stale-row counts
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                dryRun:
                  type: boolean
      operationId: graymatterLiteMemoryReindexPost
      x-thorapi-nonCrud: true
  "/memory/semantic-index/reindex":
    post:
      summary: Alias for local memory reindex
      tags:
      - LocalRuntimeOperations
      responses:
        '200':
          description: Indexed, semantic-indexed, and stale-row counts
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                dryRun:
                  type: boolean
      operationId: graymatterLiteMemorySemanticIndexReindexPost
      x-thorapi-nonCrud: true
...
components:
  schemas:
...
    GrayMatter:
      type: object
      x-thorapi-table-name: graymatter
      description: Persistent aggregate for a local GrayMatter memory workspace.
      required:
      - name
      - description
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          x-thorapi-dataField: unique=true
        description:
          type: string
          maxLength: 4000
        recommendations:
          type: string
          maxLength: 4000
        alerts:
          type: array
          maxItems: 50
          items:
            type: string
            maxLength: 500
    MemoryEntry:
      type: object
      x-thorapi-table-name: memory_entry
      description: Durable memory nucleus owned and authorized by generated ThorAPI
        services.
      required:
      - type
      - text
      properties:
        type:
          type: string
          enum:
          - configuration
          - preference
          - decision
          - todo
          - context
          - artifact
        title:
          type: string
          maxLength: 255
        text:
          type: string
          maxLength: 655350
          x-thorapi-stringStorage: portable-lob
        tags:
          type: array
          maxItems: 50
          items:
            type: string
            minLength: 1
            maxLength: 128
        sourceChannel:
          type: string
          maxLength: 128
        sourceMessageId:
          type: string
          maxLength: 256
        sourceUrl:
          type: string
          format: uri
          maxLength: 1000
        grayMatter:
          "$ref": "#/components/schemas/GrayMatter"
        retentionPolicy:
          "$ref": "#/components/schemas/MemoryRetentionPolicy"
    MemoryRetentionPolicy:
      type: object
      x-thorapi-table-name: memory_retention_policy
      description: Generated retention policy for a local memory scope.
      required:
      - name
      - retentionDays
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        retentionDays:
          type: integer
          minimum: 1
          maximum: 36500
        archiveBeforeDelete:
          type: boolean
          default: true
        description:
          type: string
          maxLength: 2000
...
    CreateMemoryEntryRequest:
      type: object
      x-thorapi-nonCrud: true
      required:
      - text
      properties:
        text:
          type: string
          maxLength: 655350
        type:
          type: string
          enum:
          - configuration
          - preference
          - decision
          - todo
          - context
          - artifact
          default: context
        sourceChannel:
          type: string
          maxLength: 128
        sourceMessageId:
          type: string
          maxLength: 256
        sourceUrl:
          type: string
          format: uri
          maxLength: 1000
        metadata:
          type: string
          maxLength: 16000
        tags:
          type: array
          maxItems: 50
          items:
            type: string
            maxLength: 128
    QueryMemoryEntryRequest:
      type: object
      x-thorapi-nonCrud: true
      properties:
        q:
          type: string
          maxLength: 4000
        query:
          type: string
          maxLength: 4000
        type:
          type: string
        source:
          type: string
          maxLength: 128
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
        maxResults:
          type: integer
          minimum: 1
          maximum: 100
    MemoryQueryResponse:
      type: object
      x-thorapi-nonCrud: true
      required:
      - results
      properties:
        results:
          type: array
          items:
            "$ref": "#/components/schemas/MemoryEntry"
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
...